move beta banner tests into header test files
