+++ features/header.feature
@@ -20,3 +20,7 @@ Feature: Home page tests
         Given I'm an sml portal user on the home page
         When I click the "glossary" link in the header
         Then The title of this page is "Library glossary"
+        
+    Scenario: Beta banner feedback link check
+        Given I'm an sml portal user
+        Then The banner mailto address is "mailto:smlhelp@ons.gov.uk"
