+++ features/steps/header.py
@@ -53,3 +53,12 @@ def check_title(context, title):
         .text
     )
     assert page_title == title
+
+@then('The banner mailto address is "{address}"')
+def check_title(context, address):
+    feedback_address = (
+        WebDriverWait(driver, timeout=timeout)
+        .until(lambda d: d.find_element(By.LINK_TEXT, "give feedback"))
+        .get_attribute("href")
+    )
+    assert feedback_address == address
